Runs without duplicates and with icons
Needs a way to remove TreeFileItems from HashTable as well as needing to reimplement zipMode name changes. Also, MenuController needs a way to select new file destination for zipMode. It needs a way to check children of directories that are blocking possible duplicates as in the branches that have been already added somewhere else with removed files. Those removed files need to be added back if that directory has been selected again. Once these goals and resolutions have been achieved, I can start to make the start button work which will make encryption or decryption start. The RSA.java file will need to be reevaluated but the core methods will be used. The next step after that would be adding functionality to the progress bar.